Only remediate ports that are enabled in config
Summary:
Only remediate ports that are enabled in config. With D5065858,
we started remediating ports via port flaps (admin disable + enable).
This should be done only for ports that are enabled in config.
Else, we will run afoul of flexports setups, where we try to lanes/ports
which cannot be enabled in a particular speed config.
For a concrete example consider a 40 or 100G quad port where lane 1
will be enabled, but lanes 2, 3, 4 are disabled. With the earlier
remediation logic we would try to enable these subports and run
afoul of flex port validation logic

Reviewed By: boryas

Differential Revision: D5106340

fbshipit-source-id: 07224230274c144a28e17d72d132bfc847f7e7d2