use updateStateNoCoalescing for port remediation
Summary:
Doing a blocking update was causing the background thread which
is the one this timeout lives on to block, which was causing issues with
the background thread not making progress, as well as slow shutdowns.
We can switch the update to be non-coalescing rather than blocking,
which should help at least unblock the background thread.

Reviewed By: aeckert

Differential Revision: D5118693

fbshipit-source-id: 34b196c5689477e45b1976f217cf73ede8c36000