Refactor RouteNextHopsMulti
Summary:
Originally, RouteNextHopsMulti is used to store client ID to nexthop set
mapping. That works for the client who sends routes with nexthops. However, for
the client to send routes with just action like DROP or ToCPU, there is no way
to track the mapping from the client ID to such info.

This diff introduces RouteNextHopEntry, which could hold either the nexthop set
like before, or the routing action. The code to use RouteNextHopEntry for
routing action is in a follow up diff.

I also created RouteNextHopBase from the original RouteNextHop. The purpose for
that is to use RouteNextHopBase to store the nexthop IP and its optional
interface ID. The next step (a separate diff) is to move RouteForwardInfo to
use RouteNextHopBase as RouteForwardInfo is almost identical to RouteNextHop.

Reviewed By: ninas

Differential Revision: D5110714

fbshipit-source-id: 98a5271302498b65be1f37c4967249ef81ee3a58