Introduce PortDescriptor object.
Summary:
This diff introduces the PortDescriptor object‐ a tagged union. A PortDescriptor is meant to act as a light-weight handle on one of several port-like entities. Its addition is motivated by the need to treat an aggregate port, which, while not a port in actuality, is nevertheless required to behave like one in most situations.

In this diff, a PortDescriptor can only act as a handle on a physical port (what we identify by a PortID). Neighbor entries and NeighborCache entries have been modified to hold a PortDescriptor in place of a PortID. In this way, the modifications herein do not alter the semantics of neighbor discovery.

Reviewed By: boryas, tfangit

Differential Revision: D4940994

fbshipit-source-id: 9c7ff3a777d7f896d997a950c37b36df645972d8