Add AggregatePort to PortDescriptor.
Summary:
This diff extends PortDescriptor to also be able to act as a handle on an AggregatePort.

New state has been added to RxPacket (ie. srcAggregatePort_ and isFromAggregatePort_) to reflect whether a packet has ingressed on an aggregate port. When an ARP/NDP packet ingresses on an aggregate port, ARP/NDP packet handlers (located in ArpHandler and IPv6Handler) use this new state to construct a PortDescriptor corresponding to the ingress aggregate port. This PortDescriptor is in turn passed through the NeighborCache machinery and subsequently stored in an ArpEntry/NdpEntry switch state node.

Lastly, SwSwitch::sendPacketOutOfPort(...) is implemented for AggregatePorts and is used in SwSwitch::sendL3Packet(...).

Reviewed By: tfangit

Differential Revision: D4948672

fbshipit-source-id: 91829ca158ea1b8f838c166b35dba300b6af0d6b