Sanitize client input during tree lookups
Summary:
If a client adds an entry to the Radix tree without properly masking the IP address, the insertion operation will properly mask it for them:

```
RadixTree<IPADDRTYPE, T, TreeTraits>::insert(...) {
  // Can't trust the clients to have 0s in all bits after mask length
  auto toAdd = ipaddr.mask(mask);
```

However, this same protection is not provided during lookup operations. As a result, a client can perform an insertion and then turn around and perform an exact match and not find anything. This diff adds the same protection (masking the IP address given by the provided masklen) to lookup operations.

Reviewed By: plapukhov

Differential Revision: D5138049

fbshipit-source-id: 0b08d558f96dbd060470cac4ad67f97bdbd9f49e