call async customize transceiver on bg thread
Summary:
Had a slight error in how we were calling the async
qsfp service calls. It is not sufficient to set the event base on
the async client to make it use that thread for IO, you also have to
actually do the call on that event base. One effect of this was that
it hit an assert when we ran a dbg build.

Reviewed By: SamanKazemkhani

Differential Revision: D5154010

fbshipit-source-id: 4b605b5381d217f9decf32ad712ac2c3342f07e6