split up port flap update into lots of non-coalescing updates
Summary:
On particularly unlucky switches, the current code hits a pretty
unpleasant snag. Suppose there are 100 bad ports (nothing plugged in on a
lab switch), then we make an update with 100 disables and an update with 100
enables. Each of those updates can take 5-10s to apply. That is mostly due
to each call to opennsl_port_speed_set taking 40ms with some taking up to 200ms.
As a result, we hold the HwSwitch mutex for up to 10s, which actually
also blocks the background thread in some neighbor cache code that accesses
the HwSwitch directly. This means that we are quite likely to be stuck
in one of these big updates when it comes time to shutdown. This can cause
~PortRemediator to time out the 5s shutdown timer, as well as a few other
places. After this change, we do the same work, but split up in lots of
small non-coalesced updates. The result of that is we don't hold that lock
for 10s and don't introduce delays on the bg or update thread. It is possible
this will result in an increase in cpu usage, though, so we should be careful.

Reviewed By: tfangit

Differential Revision: D5159981

fbshipit-source-id: cedfac698b761f615f6038013ca366d8c8d7fa29