Fix RouteNextHopsMulti [de]serialization
Summary:
Originally, RouteNextHopsMulti only has a set of IP nexthops. It has its own
serialization code. With the change to use RouteNextHopEntry within
RouteNextHopsMulti, it is the time to reuse the [de]serialization code from
other classes.

The challenge is that we need to maintain both forward and backward
compatibility so that we can warm boot into the new version. And also be
able to warm boot from new version back to old version.

This diff keeps the old [de]serialization code and also make sure the new code
will serialize to both new and old format.

Reviewed By: ninas

Differential Revision: D5157281

fbshipit-source-id: 840198f557f838a2a17bc8145c5f883b7980a5ca