Support route action per client also
Summary:
Originally, for each route prefix, we can have different clients to send
different sets of nexthops. However, there is no way for a client to specify
any route action (i.e. drop/to_cpu), as it will override the nexthop field.

This diff also slightly changed how interface route is handled. Before, a flag
is set to the route (setConnected()) when the route is created. With the new
model, setConnected() is only set if a route is resolved w/ the INTERFACE_ROUTE
client ID.

This diff also simplifies how to handle add/delete route with action or with
IP nexthops. For adding route, the API is simplified to use RouteNextHopEntry
to represent all route types (i.e. route with action, route with IP nexthops,
and interface route). For deleting a route, delRoute() is simplified to have
one version why needs client ID.

Reviewed By: ninas

Differential Revision: D5146000

fbshipit-source-id: aad13c99056901b61f6697bc0dc09d368810c341