Ability to warmboot forward and backward after action per client change
Summary:
With the latest changes to store the route Action and interface route info in
RouteNextHopsMulti now, the old assumption of RouteNextHopsMulti is empty in
such cases is broken. Because of that, the warm boot is broken between the old
code and the new code.

The diff adds function function:
1. When warm boot from old to new code, an additional logic is added to fixed
up RouteNextHopsMulti for interface and action routes.
2. During serializing, skip serializing interface route and action route into
old key, 'nexthopsmulti', so that the old code can read it back correctly. The
complete RouteNextHopsMulti is dumped to the new key, 'rib'.

Reviewed By: ninas

Differential Revision: D5170959

fbshipit-source-id: 098b727acf53fb8111f60eea815bbe3b0d146132