Remove scheduling a health check function on update EVB
Summary:
With recent FIB protection support, we could have arbitrarily
large updates pending to be applied to h/w. While in all cases that
we tested these failed updates, with thousands of routes, the update
would fail in under 30secs. However in case of a route leak, we don't
know the size of the leak in advance. And the last thing we want
is for health check to fail because of long running updates, causing
fbagent to kill fboss agent.

Instead of relying on health check function not getting scheduled we should
rely on detecting missing heartbeat on update thread. I will add a dead
detector for that which will accomplish the same thing.

Reviewed By: boryas

Differential Revision: D5174805

fbshipit-source-id: 59d17fe09934b28e16db92349e93d3cc678d54db