NAT support for DHCP relay
Summary:
On a v6 only rack, we still have v4 reachability to the TOR. In order to
support v4-only bootloader, FBOSS agent needs to support NAT for DHCP relay.

The diff adds the support to override the source IP of the DHCP relay packet to
the DHCP server, and support to override the source IP of the DHCP reply packet
to the DHCP client hosts.

The new feature is optional, completely driven by the config.

The limitation with this diff is that only one VLAN is supported. This is
because, for the DHCP reply packet, the agent code relies on the override IP to
find out the VLAN where the DHCP reply shall be forwarded out.

Reviewed By: aeckert

Differential Revision: D5064602

fbshipit-source-id: 24d6acdc9d1e1e462994db5fe6bf7b20c4c4c514