prevent segfault in IPv6Handler on agent shutdown
Summary:
TunIntf handlerReady() can result in attempting a
neighbor solicitation using the IPv6Handler object. So it isn't
safe to destroy the TunManager after the IPv6Handler as this
introduces a race. If we receive a packet from the TunIntf
after the IPv6Handler is destoryed which causes us to do a neighbor
solicitation, we will deref a null pointer and segfault.

Not sure why this doesn't happen in a widespread fashion. Maybe we
are just lucky. Noticed this happening on a fsw card in cont canary
which seemed to have more interfaces to tear down than a normal switch
so perhaps they are more susceptible to this race.

Differential Revision: D5243100

fbshipit-source-id: 227aa150d03948672c034323b55ab841ce219e6a