reduce number of calls to qsfp_service
Summary:
Due to some hacks in port change handling, we were doing
a bunch of extra work on port flaps. On disable, we would actually
hit the "changed" code path which would lead to an extra call to
program() when all we want is to do some link state stuff. Likewise, on
enable we would handle link state twice and customize twice.
Now we explicitly handle link state work in its own delta looker
atter thing which uses exactly the old criteria in processChangedPorts
but doesn't do all of program().

Reviewed By: aeckert

Differential Revision: D5253590

fbshipit-source-id: 6353bf37c954b77f5a481faadfbc8ee2aebf963d