Always update cache, not just on state changes
Summary:
Currently we only update the cached data if we catch a state change. This is bad because since we only check this when called by the agent, we could miss state changes and therefore never update the data. This means we could be making bad decisions about whether things have already been customized :(
I've also upped the timeout for cache freshness.

Reviewed By: boryas

Differential Revision: D5238824

fbshipit-source-id: ee4e9e1fafbc740c8e843b9c0178c79fa4f7d218