DROP vs ToCPU vs IP NextHops
Summary:
In the case when different path in ECMP results in differnt routing decision
(i.e. drop, to cpu, or forward to another IP), the current code will choose
DROP as the forwarding action, if DROP exists in the path.

For example,
Prefix A has 3 nexthops, NHOP_A, NHOP_B, NHOP_C.

NHOP_A is resolved by an interface route.
NHOP_B is resolved by a route w/ forwarding action DROP.
NHOP_C is resolved by a route w/ forwarding action To CPU.

The current code will choose DROP as the forwarding action of Prefix A.

The decision was based on the security consideration.

However, in the case when a default route with DROP is installed, any prefix
with originally unresolved nexthop will now be resolved as DROP due to the
default route. This causes traffic dropping.

This diff changes the way to resolve ECMP of a prefix.
- If there is at least one resovled IP path, the final forwarding info will
  include all resolved IP path
- If non path is resolved to an IP path, To CPU action will be the forwarding
  action if any of the path in ECMP is resolved to To CPU
- If all path in ECMP are resolved to DROP, the final forwarding action will be
  DROP

A new test case is added in this diff to cover the behavior above.

This diff also fixes a small return value bug in resolveMac() and one of the
test cases.

Reviewed By: ninas

Differential Revision: D5282535

fbshipit-source-id: 5c2d074a67b27bd3fe6280bba1c7c59ff8a01bb8