Extend BcmHost and BcmEgress for trunk ports.
Summary:
This diff adds a "programToTrunk" method to each of BcmHost and BcmEgress to support trunk ports.

1. BcmEgress::programToTrunk(...)

This method is different from BcmEgress::programToPort(...) in that it i) sets the OPENNSL_L3_TGID bit in opennsl_l3_egress_t.flags and uses opennsl_l3_egress_t.trunk instead of opennsl_l3_egress_t.port, and ii) does not consult the warm boot cache.

2. BcmHost::programToTrunk(...)

This method is different from BcmHost::program(...) in that it i) invokes BcmEgress::programToTrunk(...) and caches the trunk port used in the trunk_ member variable, ii) does not update the port to egress ID mapping, and iii) does not consult the warm boot cache.

Reviewed By: tfangit, boryas

Differential Revision: D4952987

fbshipit-source-id: 79bc6caf4f24f45f196385294f4a0e4022778254