minimize calls to getTransceiverInfo
Summary:
In adding flapping unexpectedly down ports, we have left
one blocking call to qsfp service. In the up half of a flap, we call setSpeed
which needs the transmitter tech (COPPER or OPTICAL). This is highly unlikely
to change so repeatedly calling getTransceiverInfo in a blocking way seems
like overkill. This diff introduces two optimizations to reduce the number
of calls to getTransceiverInfo.

The first is to simply not call
getTransmitterTech when we already have a valid value. This means we will
probably not handle switching between copper and optical on the fly properly,
which I think we are willing to do short term. On its own, that doesn't
solve the problem entirely. The reason is that ports which have nothing
plugged in will legitimately have unknown transmitter tech. There can be
lots of those (in fact they probably dominate the "down but configured up"
ports), so we need to handle them correctly. Thus, the second optimization
is to have port remediator asynchronously call getTransceiverInfo() to find
which ports are plugged in at all and only flap those which are down, enabled,
and plugged in, rather than just down and enabled.

In making that second optimization, I noticed that we should probably
use futures for creating the async client so that it for sure happens on
the background event base and in a non blocking way. That required some
light refactoring of other code that uses QsfpClient. Also, we needed access
to the platform port from PortRemediator, so I extended the Platform interface
to expose giving access to the PlatformPorts

Reviewed By: aeckert

Differential Revision: D5335032

fbshipit-source-id: c01773289018e5d110835ee607a40cfad44d33fd