destroy TunManager before NeighborUpdater
Summary:
TunManager may call in to NeighborUpdater now on receipt of a
packet. The call chain is

  TunManager::handlePacket() -> IPv6Handler::sendNeighborSolicitations() ->
  NeighborUpdater::sentNeighborSolicitation()

same possibility through the v4 version of that.

This moves destruction of TunManager earlier in the shutdown method so
we cut off all avenues for receiving packets before we start trying
to destroy other objects.

Differential Revision: D5299385

fbshipit-source-id: 3d1c2080e7e349af0d718331a9fe2352b0d63c5a