use getVia on a local evb to run getTransmitterTech in the update thread
Summary:
by switching getTransceiverInfo to being async on the bg thread then
blocking in the update context we introduced a case where the update thread
can block on the bg thread. This is unnecessary. Instead, use getVia to drive
a EventBase (so it runs in the current thread). This means we keep the nice
interface in WedgePort but let BcmPort choose to make a blocking call.
One small additional "fix" is to explicitly use via in getTransmitterTech
to avoid any possible races that could cause it to run in a diff thread.

Reviewed By: ninas

Differential Revision: D5392653

fbshipit-source-id: cefd515d957b4a60fd8557e30365eb22b0370840