stub out BcmSwitch::isHit to return true
Summary:
This is a temporary hack to make isHit always return true in
BcmSwitch. This has two benefits:

1. We no longer grab the hwlock from NeighborCache code, which can
lead to issues in the case of large updates.

2. Since we move neighbor entries from STALE to PROBE when they are
hit, we don't get cases where neighbor entries stick around forever.

This is certainly not the long term plan. Working on adding garbage
collection for stale entries that stick around forever (which is an
optional extension in the NDP RFC) to handle case 2 and am thinking
about a good way of communicating hit bit lazily to avoid needing to
grab the hw lock. However, this at least should unblock us for current
push.

Reviewed By: boryas

Differential Revision: D5403145

fbshipit-source-id: 4b0b6ba25fc6275ecbbdc0e39b6162bb75c50934