Make warmboot work with initial default route addition
Summary:
For ALPM to work we need to add v4 and v6 default routes
prior to adding any other routes to h/w. For this we added code
to add these routes to SwitchState during switch initialization
and then push them down to h/w. This works for cold boot.
However during warm boot, when there is already likely to be
a default route, RouteUpdater will return a nullptr when
applying this update (indicating nothing was added). Now if
we continue with the above code path, we will endup trying
to publish/apply a state with NULL route tables, which results
in a crash.
To fix, check result of RouteUpdater.updateDone, to see if any
routes were added, before attempting to apply switch state.

Reviewed By: boryas

Differential Revision: D5425743

fbshipit-source-id: 1b8a32e72627e8179bf88fbc0ec2abd79f83b5b1