dont look up transceiver info for ports without a transceiver id
Summary:
If a port was not configured with a transceiverID like a backplane
port on Galaxy, we were throwing an exception without registering any future
cobs or vias so onError or inspecting the Try passed into a callback weren't
handling the problem. Moving the dangerous call into the callbacks makes it
so that we can catch those errors if we care to. Otherwise, those transceivers
are skipped.

Reviewed By: aeckert

Differential Revision: D5443217

fbshipit-source-id: 0f49f09c8947191f50e7ba7d7840d710c72ec2f4