remove redundant getPortSpeed/getMaxPortSpeed calls in HwSwitch interface
Summary:
We do some pretty circuitous stuff on startup for populating
the speed + maxspeed fields on port switchstate nodes. In the cold
boot case, we return ports w/ DEFAULT speed and then set the speed
from SwSwitch::init. In warm boot, this is done in
BcmSwitch::getWarmBootSwitchState. Because of this logic, we have
getPortSpeed and getMaxPortSpeed calls in the main HwSwitch interface,
which is completely redundant since we also have the exact same
information in our SwitchState.

This removes these two calls from the HwSwitch interface, sets speed
and maxSpeed in BcmSwitch::getColdBootSwitchState so it is always set,
and hits new BcmPort::getMaxSpeed() api so that the max speed
population works w/ new-style flexports, which fixes an issue we are
currently seeing on galaxy.

Reviewed By: ninas

Differential Revision: D5443580

fbshipit-source-id: cbf107eb426440de3c332fbf5206217539df4915