add back accidentally removed max port speed validation
Summary:
I started removing this logic since it is not completely
necessary, but ended up fixing getting the max speed instead. However,
I had still deleted these lines in D5443580. This adds them back
in. It is not strictly necessary and we are unlikely to hit it, but it
gives useful error messages in some cases.

Reviewed By: boryas

Differential Revision: D5452216

fbshipit-source-id: 5d4611f440740e198c9c9c65135023648e05c37b