dont disable linkscan on each port disable
Summary:
In the port disable path, we disable linkscan for the port. This is
a legitimate part of correctly downing a port when reconfiguring lanes during
a speed change with flexports. However, it leads to not getting a linkscan
callback when we disable a port. Previously, we got lucky and some part of
"setSpeed" either legitimately or incorrectly triggered a linkscan callback
even after we disabled linkscan. However, we shouldn't actually call setSpeed
on disable. When we removed that error in D5253590 it exposed this bug.

The fix is to do the linkscan disable/enable only during lane reconfiguration
rather than on every enable/disable.

Reviewed By: aeckert

Differential Revision: D5450971

fbshipit-source-id: 45faa7c8bf1c94954f28193802a4f19ecf844c94