BcmPort::getState() --> BcmPort::isUp()
Summary:
Next step in the crusade against confusing port states is to
get rid of the last known (to me) usage of PortState::DOWN. It was a
possible return value from BcmPort::getState() which indicated that the
port is administratively up but operationally down. Fortunately, all the
callsites were just checking if it was equal to PortState::UP, so changing
it to 'isUp' was pretty natural.

Reviewed By: aeckert

Differential Revision: D5459514

fbshipit-source-id: 3c2dd44e7c7036e273d3fa7655752c1ec822fbef