add non-fatal callback for parity error switch events
Summary:
We recently registered a handler for parity errors in order
to count them. However, we used the
BcmSwitchEventUnitFatalErrorCallback class, which will always fatal
and exit the agent. Parity errors are not necessarily unrecoverable
so we should not just crash when this happens. There are about 80
switches hitting parity errors right now.

Diff adds a new class for logging non-fatal errors and registers that
one on parity errors.

Reviewed By: boryas

Differential Revision: D5491324

fbshipit-source-id: 6c49208b7e70cc738a25a6ee749f00286b2e4b9f