Try2: Pull the revision number from fabricgen generated config
Summary:
Let's keep the version type as integer. We don't actually need to change it. So for now agent version is an integer and bgpd policy version is a string. We can easily cast them as integer while calling and doing the calculations in audit logic.

Details in T17234949

- Pull the fabricgen config revision number and fill it as agent & bgpd config version.
- For RSW set it to 0, for FSW use the values from materialized JSON
- Remove the SixPack unit tests (They were showing as failed as the corresponding test data has been removed)

Last iteration: D5426207

Differential Revision: D5542656

fbshipit-source-id: 8a951747e69dc4fc875f696091d5eb99aad9fbce