change port flaps to customize on bg thread
Summary:
We have reason to believe that the high amount of flapping we
are doing could be causing problems. We initially switched to flapping as a
way to accomplish customizing transceivers in a nice synchronized way (via
update thread). While this is not perfect (PlatformPort is not thread safe)
since WedgePort is essentially static and we serialize the customize
calls on the background thread, this ought to be safe.

Differential Revision: D5608661

fbshipit-source-id: 230193e3b010476505fe8f10b8ebeae59bf22a1c