move fully to AdminState/OperState model
Summary:
This diff deprecates PortState::DOWN and does the renaming
to start calling state adminState. It also makes operState an enum for
consistency. It also adds some convenience methods to Port for
looking into the states. As a lifetime hater of double negatives, I swapped
isAdminDisabled for isEnabled. isUp handles operState. Almost every callsite
of isAdminDisabled negated it, so I think this is an improvement for
readability.
One major thing to note is the convoluted upgrade path for to/fromFollyDynamic.
Renaming the enum values is sort of tricky and will require two releases to
fully do and get rid of these hacky handmade maps.

Reviewed By: aeckert

Differential Revision: D5476496

fbshipit-source-id: 103a38d7e7c9700f5196132293b99a2347818beb