Introduce per-member forwarding state to AggregatePort.
Summary:
Previously, every (physical) member port of an aggregate port would be used to egress out of the aggregate port. This diff allows an aggregate port to use only a subset of all of its member ports for forwarding.

This is accomplished by extending AggregatePort to include a per-member, binary state having values Forwarding::ENABLED and Forwarding::DISABLED.

The per-member, forwarding state will be used 1) in link-down handling, to avoid black-holing traffic by forwarding packets out of down member ports, 2) in link-up handling, to avoid forwarding traffic over a link whose PHY is not yet usable, and from the 3) Link Aggregation Control Protocol, to signal down into the SwitchState the physical ports which the local system and the remote system have agreed to aggregate.

It's worthwhile keeping the following distinction in mind: the Forwarding state added herein is a run-time property of member ports (eg. it is altered by link events and LACP), whereas the subports of each aggregate port is a configuration-time property.

Depends on D5432484.

Differential Revision: D5414350

fbshipit-source-id: b7b86bc3f85957d24b2ccdab7bb1131e4ad7da7f