Maintain a mapping from trunk ports to egress ids.
Summary:
To avoid blackholing traffic in the case of a link-down event, all egresses which use the newly-downed port are found and subsequently removed from all ECMP entries. To support link-down handling for aggregate ports, this diff extends PortAndEgressIdsMap to also hold trunk egresses.

This is accomplished by modifying the PortAndEgressIdsMap's key to be a gport, which can represent either a trunk port or a physical port (in addition to many other flavors of ports).

Depends on D5414350.

Differential Revision: D5506176

fbshipit-source-id: a599ecafaea4fe64e67f1c8d36fd45caab03cca8