FBOSS] port pause config HW programming
Summary:
Add the support to program HW for the Tx/Rx pause setting.

As part of the change, I also removed the code that always disable pause when
port is created. Now, the pause setting is controlled by the config.

Before this change, all ports, no matter enabled or not, had pause disabled.
With this change, only the ports that are enabled will have correct pause
configured based on the configuration. However, since this difference is only
on the ports that are not enabled, there shall not have any side effect to the
switch.

For warm boot case, it is protected by both the switch config (no change during
switch config), and the get/compare/set-if-diff logic in the setPause()
function.

Reviewed By: aeckert

Differential Revision: D5618971

fbshipit-source-id: 6094045ab3835c829df90f2e2ec9c58bb759875e