Fix thrift dependencies
Summary: Correct dependencies to use cmake. Note the current cmake of thrift can't handle async builds. The reason is that the libraries required generated files. Even when specifying dependencies on libraries that are autogenerated the async build does not correctly respect this. There should be a hack around this but for now this will build thrift synchronously.

Reviewed By: capveg

Differential Revision: D5747726

fbshipit-source-id: c26b742b09f4e1be4266fcea86bfca5b00f956d1