Add api for returning raw DOM page data
Summary:
This adds a new thrift endpoint to just retrieve the raw DOM
data for various transceivers. The use case is that we can
programatically inspect QSFP DOM data for fields that may not yet have
a representation in our thrift structs.

An example of this is I wanted to get a sense of transceiver part
numbers that incorrectly implement the fractional length field we ask
of our vendors. To do this at scale I either need to use qsfp_util
(which contends w/ qsfp_service for i2c bus access), or I need to roll
a new qsfp_service.

Adding a raw read api makes it so we can query DOM data ahead of
implementing full parsing logic. We can also change the default of
wedge_qsfp_util to hit this api in order to be a bit safer.

Reviewed By: tfangit

Differential Revision: D5656770

fbshipit-source-id: 3c66a5d969e0f021be8f8b4e4d1b4dc2d8a9ed93