properly set channels in getPortStatus
Summary:
In ctrl.thrift we have had TransceiverIdx.channelId marked
'deprecated' since 2015. The reason is that a port can actually be
made up of multiple channels in a transceiver so saying it has only
one channel is a bit misleading. The cli has some hacky logic for
backwards compatibility, but it looks like we never actually pushed
the agent changes needed to set the TransceiverIdx.channels field
properly.

This changes fillPortStatus to properly set the channels field and no
longer set the channelId field.

This is a bit worrisome since we access the BcmPort directly through
the BcmPort* stored in the platform port, which could certainly race
w/ the update thread on changes. We never destroy BcmPort or
BcmPortGroup objects in normal operation though so this should be
alright. It would be better to model the PortGroup somehow in
SwitchState to avoid races like this.

Differential Revision: D5686161

fbshipit-source-id: 4c5b85e12de3fa9dc97c32c810337de82ff68ab9