throw exception instead of segfaulting on tx pkt allocation failure
Summary:
We currently log an "Out of memory" message when we fail to
allocate dma memory for a packet. However, we then segfault on the
next line trying to access pkt_ . This changes logic to throw a more
detailed exception instead of just segfaulting, so we can potentially
do some error handling as well before exiting.

Some callsites will catch the exception (eg TunIntf.cpp), while others
will not (IPv6Handler::sendNeighborSolicitation). We should
standardize on what the best way to handle this is. Sending this out
because segfaulting is certainly not the best solution.

Reviewed By: plapukhov

Differential Revision: D5734198

fbshipit-source-id: ed403d7145dd89f8d46b585181730fc116d466a9