Rename link_state.down to link_state.flap
Summary:
Right now, we have link_state.down for switch as a whole and each port to keep track of the number of flaps happens
in a period of time. Instead of using `down  as a key, we should use `flap` to be more accurate. Besides, I also fix
the key name for a single port with a dot to separate port# and link_state to make it more readable.

To address jasmeetbagga comment about using ethX/Y/Z format to make it more readable. I have to introduce the portName into PortStats.
However that also brings some potential difficulties for SwitchStats/PortStats:
1) This is the first counter key logged into ODS in s/w side, unlike h/w BcmPort already has its `portName_` attribute, to get portName from portID, we need to depend on SwitchState from SwSwitch.
2) How can we recognize portName change in s/w side?
3) In SwSwitch, we decided to use folly::ThreadLocalPtr<SwitchStats, SwSwitch> stats_ to maintain per SwitchStats per thread in agent. Therefore, if one thread notice there's a new port created/ old port deleted/ port name change, we need to update all PortStats for all running thread

Here are very general approaches to overcome those difficulties above:
1) Keep SwitchStats/PortStats independent from SwitchState. To do that the new PortStats need to explicitly pass portName in its constructor, and when StateStats is created in SwSwitch, it will also generate PortStats for every port in current state.
2) Like other StateObserver, I created a new PortHandler extends AutoRegisterStateObserver to update PortStats if there's a port update. The previous code in SwSwitch to update the port up/down status counter is also moved to PortStats and the `setPortStatusCounter` will be called if there's a port update. Now we keep the same functionality and make SwSwitch cleaner.
3) Calling stats_.accessAllThreads() to get every thread SwitchStats and then update its PortStats if needed.

Differential Revision: D5695076

fbshipit-source-id: 376c2ec4da272c1706b63abc6e88c279fbc1a42e