Adding CLI "fboss nic vendor" with "--detail".
Summary:
This is a stub for adding host NIC information into fboss CLI.
For now, this returns the vendor name on issuing "fboss nic vendor" from CLI.
The "--detail" option provides the NIC of the vendor per MAC address.

The need for this is seen, especially when we were debugging issues with FEC.

This can possibly be expanded in future to have:
  fboss nic buffer
  fboss nic configuration
  fboss nic details
  fboss nic fec <>
  fboss nic speed
  fboss nic vendor
Some more low priority possibilities are --avoid_uplink or --avoid_port or --avoid_nic.

Reviewed By: capveg

Differential Revision: D5735150

fbshipit-source-id: 0eefed3b7b458838bfec54896cfb89e835fd2fab