start adding HwTestHandle
Summary:
This refactors our tests to instantiate a HwTestHandle that
owns the SwSwitch and the Mockable platform. The HwTestHandle
implements a few functions, (rxPacket, forcePortDown, forcePortUp,
forcePortFlap), but we can extend this as needed for other tests.

The initial construction logic + api is a bit awkward and there is a
decent amount of copypasta around our tests (though that was the case
already). My plan is to refactor the construction in to a few base
TestFixtures that our tests can extend. Definitely open to suggestions
about improving the layering of these apis.

For now, every platform uses the HwTestHandle defined in MockPlatform,
which should match the current behaviour of the tests. Will send out
subsequent diffs that add HwTestHandle implementations for BcmPlatform
and FakeSdk.

Reviewed By: plapukhov

Differential Revision: D4772253

fbshipit-source-id: ad50b6b5e2be878b22bc2258f1e7b1d4894fb4b4