NDPTest TSAN bugs
Summary:
This starts fixing some bugs TSAN uncovered in our
NDPTest. Basically we had broken logic where we would do the following:

1. take action X
2. wait for state updates
3. get and inspect state

There is a pretty obvious race between 2/3 if you insert an arbitrary
sleep, because the state may change in that time. This diff changes
the waitForStateUpdates api to actually return the resulting state at
the point that the NoOp state update is processed so that we inspect
the correct state generation regardless of thread scheduling.

Reviewed By: boryas

Differential Revision: D4772373

fbshipit-source-id: 67ed4dd11eee1384c52816d8da83e0e89a9d801c