Retrieve egressID from warmboot file instead of HW host table
Summary:
Originally, we retrieve all egressID by reading entries from the HW host table.
This is based on the assumption that all BcmHosts are programmed to HW, which
is true.

In order to support v6 link-local as the nexthop, we are changing to not
program v6 link-local BcmHost to the HW. In this case, we cannot retrieve all
egressIDs pointed by the BcmHost from HW during warmboot.

This diff changes to use warm boot file instead of HW to retrieve all egressIDs
pointed to BcmHost. We will need to land this first. Only after that, we can
support warmboot forward and backward in the future when supporting v6
link-local nexthop.

Reviewed By: boryas

Differential Revision: D5764831

fbshipit-source-id: cc356598b18936b3cc5232723528c7a60b7f2658