Remove a BcmHost constructor variance
Summary:
There is a requirement to create BcmHost object for route to be installed in
the host table. In this case, egress ID is provided. A 2nd constructor was
created to pass in the egress ID.

This diff merges two constructors back together and introduced a method just
for set egress ID. The change also helps to remove variable template parameters
for incRefOrCreateHcmHostImpl() func.

Differential Revision: D5764837

fbshipit-source-id: 3905a5f4eb7fed8f7337343a844083382d11289d