Don't fill up disc on non fatal ASIC errors
Summary:
We have had cases where parity errors cause us to
fill up the disc and cause other instabilities in the system.
We already log these errors to ODS. In case of persistent
parity errros we further raise alerts and tasks for a human
to intervene.
Further these parity errors are correctable and switch can
(mostly) continue forwarding in their presence. So no need
to fill up disc and make other parts of the system unstable

Reviewed By: boryas

Differential Revision: D5819670

fbshipit-source-id: 672920dc96fb0ae8c43eda0d9fd63bdde4bcb71a