Remove pcap_distrbutor deconstructor race
Summary:
pcap_distributor had a race condition when closing its channel.
If the channel was closed before the destuctor was called,
everything exitted fine.  But, if it was still active when
the destructor was called, the closer_ ojbect was destructed
before pcap_distributor, and thus filed with garbage.

re-ordered to enforce the proper destruction order

Reviewed By: boryas

Differential Revision: D5847044

fbshipit-source-id: c049fada79d203a6c7685c91cb1e154923cdf353